Refactor Scoring panel
