+++ src/dashboard/scoring/colors.ts
@@ -1,0 +1,128 @@
+import { ColorGroup } from 'bundle/types/colorGroup';
+
+export const colors: ColorGroup[] = [
+    {
+        meta: {
+            name: 'Ranked Modes'
+        },
+        colors: [
+            {
+                index: 0,
+                title: 'Green vs Grape',
+                clrA: '#37FC00',
+                clrB: '#7D28FC'
+            },
+            {
+                index: 1,
+                title: 'Green vs Magenta',
+                clrA: '#04D976',
+                clrB: '#D600AB'
+            },
+            {
+                index: 2,
+                title: 'Turquoise vs Orange',
+                clrA: '#10E38F',
+                clrB: '#FB7B08'
+            },
+            {
+                index: 3,
+                title: 'Mustard vs Purple',
+                clrA: '#FF9E03',
+                clrB: '#B909E0'
+            },
+            {
+                index: 4,
+                title: 'Dark Blue vs Green',
+                clrA: '#2F27CC',
+                clrB: '#37FC00'
+            },
+            {
+                index: 5,
+                title: 'Purple vs Green',
+                clrA: '#B909E0',
+                clrB: '#37FC00'
+            },
+            {
+                index: 6,
+                title: 'Yellow vs Blue',
+                clrA: '#FEF232',
+                clrB: '#2ED2FE'
+            }
+        ]
+    },
+    {
+        meta: {
+            name: 'Turf War'
+        },
+        colors: [
+            {
+                index: 8,
+                title: 'Yellow vs Purple',
+                clrA: '#D1E004',
+                clrB: '#960CB2'
+            },
+            {
+                index: 9,
+                title: 'Pink vs Blue',
+                clrA: '#E61077',
+                clrB: '#361CB8'
+            },
+            {
+                index: 10,
+                title: 'Pink vs Yellow',
+                clrA: '#ED0C6A',
+                clrB: '#D5E802'
+            },
+            {
+                index: 11,
+                title: 'Purple vs Turquoise',
+                clrA: '#6B10CC',
+                clrB: '#08CC81'
+            },
+            {
+                index: 12,
+                title: 'Pink vs Light Blue',
+                clrA: '#E30960',
+                clrB: '#02ADCF'
+            },
+            {
+                index: 13,
+                title: 'Purple vs Orange',
+                clrA: '#5617C2',
+                clrB: '#FF5F03'
+            },
+            {
+                index: 14,
+                title: 'Pink vs Green',
+                clrA: '#E60572',
+                clrB: '#1BBF0F'
+            }
+        ]
+    },
+    {
+        meta: {
+            name: 'Color Lock'
+        },
+        colors: [
+            {
+                index: 7,
+                title: 'Yellow vs Blue (Color Lock)',
+                clrA: '#FEF232',
+                clrB: '#2F27CC'
+            }
+        ]
+    },
+    {
+        meta: {
+            name: 'Custom Color'
+        },
+        colors: [
+            {
+                index: 999,
+                title: 'Custom Color',
+                clrA: '#000000',
+                clrB: '#FFFFFF'
+            }
+        ]
+    }
+];
