+++ src/dashboard/scoring/main.ts
@@ -1,0 +1,9 @@
+import { NodeCGBrowser } from 'nodecg/browser';
+
+import './scores';
+import './scoreboardData';
+import './colorData';
+import './nextTeams';
+
+import '../globalStyles.css';
+import './scoring.css';
