+++ src/types/colorGroup.ts
@@ -1,0 +1,6 @@
+import { ColorInfo } from './colorInfo';
+
+export interface ColorGroup {
+    meta: { name: string },
+    colors: ColorInfo[]
+}
