+++ src/types/colorInfo.ts
@@ -1,0 +1,6 @@
+export interface ColorInfo {
+    index: number,
+    title: string,
+    clrA: string,
+    clrB: string
+}
